Add --lib flag for specifying an image as library
This gives the vm (and therefore any command that runs the VM) a flag
`--lib`, which will cache and use the mentioned image in the module
search path.

At present the module search path only uses a FileResolver; that is,
it expects modules in a flat structure.

Ths involves a bit of extra test scaffolding, so that invocations of
`jk` when testing can use `--lib` (and `--cache` to set the cache to a
fresh, temporary directory). In particular, an image registry is
started, and images from tarballs uploaded to it. This is described in
`tests/testfiles/README-images.md`.