Account for whiteout files
In the OCI image format, the layer filesystems can contain "whiteout"
files (think liquid paper or tipp-ex). This suppress any like-named
file in the layers below.

    https://github.com/opencontainers/image-spec/blob/master/layer.md#whiteouts

This commit accounts for whiteouts in the overlay VFS when opening a
file or enumerating files in a directory.

It's a fairly naive implementation, which checks each possible
whiteout file, when traversing layers (since whiteouts hide things in
the _next_ layer).